+++ infra/modules/gcp-psoxy-rest/main.tf
@@ -187,7 +187,7 @@ resource "local_file" "test_script" {
   count = var.todos_as_local_files ? 1 : 0
 
   filename        = "test-${trimprefix(var.instance_id, var.environment_id_prefix)}.sh"
-  file_permission = "0770"
+  file_permission = "755"
   content         = <<EOT
 #!/bin/bash
 API_PATH=$${1:-${try(var.example_api_calls[0], "")}}