+++ java/core/src/main/java/co/worklytics/psoxy/PseudonymizerImpl.java
@@ -102,7 +102,7 @@ public class PseudonymizerImpl implements Pseudonymizer {
             scope = getOptions().getDefaultScopeId();
         }
 
-        builder.scope(scope);
+        builder.scope(StringUtils.trimToNull(scope));
 
 
         byte[] hashWithDefaultImpl =