+++ java/impl/aws/src/main/java/co/worklytics/psoxy/S3Handler.java
@@ -72,19 +72,21 @@ public class S3Handler implements com.amazonaws.services.lambda.runtime.RequestH
             return null;
         }
 
-        S3Object s3Object = s3Client.getObject(new GetObjectRequest(importBucket, sourceKey));
+        S3Object sourceObject = s3Client.getObject(new GetObjectRequest(importBucket, sourceKey));
         byte[] processedData = null;
-        try (InputStream objectData = s3Object.getObjectContent();
-             BOMInputStream is = new BOMInputStream(objectData);
-             InputStreamReader isr = new InputStreamReader(is, StandardCharsets.UTF_8);
-             Reader reader = new BufferedReader(isr);
-             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-             OutputStreamWriter outputStreamWriter = new OutputStreamWriter(outputStream);
-             Writer writer = new BufferedWriter(outputStreamWriter)) {
+        try (InputStream objectData = sourceObject.getObjectContent();
+             ByteArrayOutputStream outputStream = new ByteArrayOutputStream()) {
 
-            StorageEventRequest request = storageHandler.buildRequest(reader, writer, importBucket, sourceKey, transform);
+            StorageEventRequest request = storageHandler.buildRequest(null, null, importBucket, sourceKey, transform);
 
-            storageEventResponse = storageHandler.handle(request, transform.getRules());
+            boolean isCompressed = Optional.ofNullable(sourceObject.getObjectMetadata().getContentEncoding())
+                .map(s -> s.contains("gzip"))
+                .orElse(false);
+
+            request = request.withDecompressInput(isCompressed).withCompressOutput(isCompressed);
+
+            storageEventResponse = storageHandler.process(request, transform, objectData, outputStream);
+
             processedData = outputStream.toByteArray();
             log.info(String.format("Successfully pseudonymized %s/%s to buffer", importBucket, sourceKey));
         }
@@ -93,7 +95,14 @@ public class S3Handler implements com.amazonaws.services.lambda.runtime.RequestH
             ObjectMetadata meta = new ObjectMetadata();
             //NOTE: not setting content length here causes S3 client to buffer the stream ... OK
             //meta.setContentLength(storageEventResponse.getBytes().length);
-            meta.setContentType(s3Object.getObjectMetadata().getContentType());
+
+            // set headers iff they're non-null on source object
+            Optional.ofNullable(sourceObject.getObjectMetadata().getContentType())
+                .ifPresent(meta::setContentType);
+            Optional.ofNullable(sourceObject.getObjectMetadata().getContentEncoding())
+                .ifPresent(meta::setContentEncoding);
+
+
             meta.setUserMetadata(storageHandler.buildObjectMetadata(importBucket, sourceKey, transform));